Level: Parametrise the Player, Goober, and Explosion scenes
Previously, the Level script had hardcoded references to the default
versions of these scenes. This means that (for example) even if you use a
TileMapLayer with a different TileSet that displays the player with a
different sprite, the default Player appearance would be used.

With this change, the Player could be customised for a given world (or
even a single level) by duplicating the Player scene, overriding its
sprites (but keeping the script the same), and then setting the
player_scene property on the level.